requirements.txt
aiohttp==3.6.3
aiostream==0.4.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
braceexpand==0.1.6
chardet==3.0.4
dateparser==1.0.0
discord==1.0.1
discord-ext-forms==2.9.1
discord-ext-wizards==0.3.0
discord-pretty-help==1.2.1
discord.py==1.7.2
disputils @ git+https://github.com/LiBa001/disputils@13fd04e88dac541a0964497139527e6aa9246fe0
dnspython==2.1.0
emoji==1.2.0
humanize==3.0.1
idna==2.10
import-expression==1.1.4
jishaku==1.20.0.220
marshmallow==3.8.0
motor==2.3.0
multidict==4.7.6
natsort==7.1.1
normal-api.py==1.0.0
pymongo==3.11.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2021.1
regex==2020.11.13
six==1.15.0
typing-extensions==3.7.4.3
tzlocal==2.1
umongo==3.0.0b10
yarl==1.5.1
